server/requirements.txt
aerich~=0.7.0
bcrypt~=4.0.0
PyJWT~=2.6.0
tornado~=6.3.3
tornado-middleware~=1.2.3
tortoise-orm~=0.20.0
